requirements.txt
allure-pytest==2.7.1
allure-python-commons==2.7.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
lazy-object-proxy==1.4.1
loguru==0.3.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.2
pytest==5.0.1
pytest-html==1.22.0
pytest-metadata==1.8.0
python-dotenv==0.10.3
requests==2.22.0
selenium==3.141.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
